fix: prevent int overflow when calculating range size (#37)
Update the code calculating the size of a range to use a `long` variable
instead of an `int` to avoid overflow errors when selecting extremely large
ranges (e.g. an entire sheet).
To ensure consistency, update all variables dealing with range sizes to use
a `long` type instead of `int`.

BREAKING CHANGE: All references to range size now stored as long variables instead of int

+++ ExcelChangeHandler/ChangeHandling/Memory/ChangePropertiesImpl.cs
@@ -12,11 +12,11 @@ namespace ExcelChangeHandler.ChangeHandling.Memory
 
         public string? RangeAddress { get; }
 
-        public int? RangeCellCount { get; }
+        public long? RangeCellCount { get; }
 
         public string[,]? RangeFormulas { get; }
 
-        public ChangePropertiesImpl(string? sheetName, int? sheetColumns, int? sheetRows, string? rangeAddress, int? rangeCellCount, string[,]? rangeFormulas)
+        public ChangePropertiesImpl(string? sheetName, int? sheetColumns, int? sheetRows, string? rangeAddress, long? rangeCellCount, string[,]? rangeFormulas)
         {
             SheetName = sheetName;
             SheetColumns = sheetColumns;
